requirements.txt
flask
numpy
Pillow
tensorflow
